scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.2
pyyaml==6.0
tqdm==4.66.1
tenacity==8.2.2
lxml==4.9.3
validators==0.21.0
openpyxl==3.1.2
